+++ CHANGELOG.md
@@ -2,6 +2,8 @@
 
 * fix: hide "Shared reference" entry from menu if
   shared data is not initialized
+* Hermine reads version from pyproject.toml and display it in the footer
+
 
 # 0.0.1